refactor(post-manager-service): rebranding service to dialosoft again
