+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/mapper/ReactionsMapper.java
@@ -1,0 +1,11 @@
+package com.dialosoft.postmanager.mapper;
+
+import com.dialosoft.postmanager.models.entities.ReactionsEntity;
+import com.dialosoft.postmanager.models.web.request.CreateReactionCommonAttributes;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ReactionsMapper {
+
+    ReactionsEntity toEntity(CreateReactionCommonAttributes source);
+}
