+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/dto/BasicCommonAttributes.java
@@ -1,0 +1,15 @@
+package com.dialosoft.postmanager.models.dto;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class BasicCommonAttributes {
+    String id;
+    String username;
+    String content;
+}
