+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/dto/Reactions.java
@@ -1,0 +1,17 @@
+package com.dialosoft.postmanager.models.dto;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class Reactions {
+    String postId;
+    String commentId;
+    int countReactions;
+    int positiveReaction;
+    int negativeReaction;
+}
