+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/web/request/CreateReactionCommonAttributes.java
@@ -1,0 +1,16 @@
+package com.dialosoft.postmanager.models.web.request;
+
+import com.dialosoft.postmanager.models.dto.BasicCommonAttributes;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class CreateReactionCommonAttributes extends BasicCommonAttributes {
+    Boolean reaction;
+    String postId;
+    String commentId;
+}
