+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/web/request/PostManagerCommonAttributes.java
@@ -1,0 +1,14 @@
+package com.dialosoft.postmanager.models.web.request;
+
+import com.dialosoft.postmanager.models.dto.BasicCommonAttributes;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class PostManagerCommonAttributes extends BasicCommonAttributes {
+    String image;
+}
