--- src/post-manager/src/main/java/com/biznetbb/postmanager/mapper/ReactionsMapper.java
@@ -1,11 +1,0 @@
-package com.biznetbb.postmanager.mapper;
-
-import com.biznetbb.postmanager.models.entities.ReactionsEntity;
-import com.biznetbb.postmanager.models.web.request.CreateReactionCommonAttributes;
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface ReactionsMapper {
-
-    ReactionsEntity toEntity(CreateReactionCommonAttributes source);
-}
