--- src/post-manager/src/main/java/com/biznetbb/postmanager/models/dto/Comments.java
@@ -1,18 +1,0 @@
-package com.biznetbb.postmanager.models.dto;
-
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class Comments extends BasicCommonAttributes{
-    Integer positiveReaction;
-    Integer negativeReaction;
-    LocalDateTime creationTime;
-}
