--- src/post-manager/src/main/java/com/biznetbb/postmanager/models/web/request/CreateCommentCommonAttributes.java
@@ -1,15 +1,0 @@
-package com.biznetbb.postmanager.models.web.request;
-
-import com.biznetbb.postmanager.models.dto.BasicCommonAttributes;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class CreateCommentCommonAttributes extends BasicCommonAttributes {
-    String commentId;
-    String postId;
-}
