--- src/post-manager/src/main/java/com/biznetbb/postmanager/repository/PostManagerRepository.java
@@ -1,13 +1,0 @@
-package com.biznetbb.postmanager.repository;
-
-import com.biznetbb.postmanager.models.entities.PostEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-public interface PostManagerRepository extends JpaRepository<PostEntity, UUID> {
-    // remember using optinal for the methods of this repository
-
-    Optional<List<PostEntity>> findByUsername(String username);
-}
